requirements.txt
flask==2.2.3
werkzeug==2.2.3
tensorflow==2.13.0
Pillow==9.4.0
numpy==1.23.5
uuid==1.30
requests==2.28.2 